feat: Add basic authentication
