+++ examples/consumer.rs
@@ -4,8 +4,8 @@ use std::env;
 
 use futures::TryStreamExt;
 use pulsar::{
-    authentication::oauth2::OAuth2Authentication, Authentication, Consumer, DeserializeMessage,
-    Payload, Pulsar, SubType, TokioExecutor,
+    authentication::{basic::BasicAuthentication, oauth2::OAuth2Authentication},
+    Authentication, Consumer, DeserializeMessage, Payload, Pulsar, SubType, TokioExecutor,
 };
 
 #[derive(Serialize, Deserialize)]
@@ -46,6 +46,11 @@ async fn main() -> Result<(), pulsar::Error> {
             serde_json::from_str(oauth2_cfg.as_str())
                 .unwrap_or_else(|_| panic!("invalid oauth2 config [{}]", oauth2_cfg.as_str())),
         ));
+    } else if let (Ok(username), Ok(password)) = (
+        env::var("PULSAR_BASIC_USERNAME"),
+        env::var("PULSAR_BASIC_PASSWORD"),
+    ) {
+        builder = builder.with_auth_provider(BasicAuthentication::new(&username, &password))
     }
 
     let pulsar: Pulsar<_> = builder.build().await?;
